requirements.txt
phidata
sqlalchemy
mysql-connector-python
groq
Flask
elevenlabs
openai

-e .